ongoing bug fixes and improvement branch for pill prox testing.

+++ drivers/prox_i2c.h
@@ -2,4 +2,4 @@
 #include "message_base.h"
 
 MSG_Status init_prox(void);
-void read_prox(uint16_t * out_val1, uint16_t * out_val4);
+void read_prox(uint32_t * out_val1, uint32_t * out_val4);
